move PersistentSubscription-* classes out of internal folder + namespace
