+++ src/Async/EventStoreConnection.php
@@ -22,11 +22,11 @@ use Prooph\EventStore\DeleteResult;
 use Prooph\EventStore\EventData;
 use Prooph\EventStore\EventReadResult;
 use Prooph\EventStore\EventStoreSubscription;
-use Prooph\EventStore\Internal\PersistentSubscriptionCreateResult;
-use Prooph\EventStore\Internal\PersistentSubscriptionDeleteResult;
-use Prooph\EventStore\Internal\PersistentSubscriptionUpdateResult;
 use Prooph\EventStore\ListenerHandler;
+use Prooph\EventStore\PersistentSubscriptionCreateResult;
+use Prooph\EventStore\PersistentSubscriptionDeleteResult;
 use Prooph\EventStore\PersistentSubscriptionSettings;
+use Prooph\EventStore\PersistentSubscriptionUpdateResult;
 use Prooph\EventStore\Position;
 use Prooph\EventStore\RawStreamMetadataResult;
 use Prooph\EventStore\ResolvedEvent;
